MET-4284 change data type to support null in total amount of records

+++ src/main/java/eu/europeana/metis/sandbox/service/dataset/DatasetServiceImpl.java
@@ -74,7 +74,7 @@ class DatasetServiceImpl implements DatasetService {
 
   @Override
   @Transactional
-  public void updateNumberOfTotalRecord(String datasetId, int numberOfRecords) {
+  public void updateNumberOfTotalRecord(String datasetId, Long numberOfRecords) {
     datasetRepository.updateRecordsQuantity(Integer.parseInt(datasetId), numberOfRecords);
   }
